requirements.txt
Django~=3.2.5
django-autoslug~=1.9.8
Pillow==8.3.1
